backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.5
django-environ==0.9.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
